Moved TestUserProvider into /src/testing
