+++ frontend/src/app/DevHomePage.test.tsx
@@ -1,9 +1,9 @@
 import { beforeEach, describe, expect, test } from "vitest";
 
-import { TestUserProvider } from "@/hooks/user/TestUserProvider";
 import { ElectionListRequestHandler } from "@/testing/api-mocks/RequestHandlers";
 import { server } from "@/testing/server";
 import { render, screen } from "@/testing/test-utils";
+import { TestUserProvider } from "@/testing/TestUserProvider";
 
 import { DevHomePage } from "./DevHomePage";
