+++ frontend/src/components/navbar/NavBar.stories.tsx
@@ -3,8 +3,8 @@ import * as React from "react";
 import { Story } from "@ladle/react";
 
 import { ElectionProviderContext } from "@/hooks/election/ElectionProviderContext";
-import { TestUserProvider } from "@/hooks/user/TestUserProvider";
 import { electionDetailsMockResponse } from "@/testing/api-mocks/ElectionMockData";
+import { TestUserProvider } from "@/testing/TestUserProvider";
 import { Election, Role } from "@/types/generated/openapi";
 
 import { NavBar } from "./NavBar";
