+++ frontend/src/components/navbar/NavBar.test.tsx
@@ -2,10 +2,10 @@ import { userEvent } from "@testing-library/user-event";
 import { beforeEach, describe, expect, test } from "vitest";
 
 import { ElectionProvider } from "@/hooks/election/ElectionProvider";
-import { TestUserProvider } from "@/hooks/user/TestUserProvider";
 import { ElectionRequestHandler } from "@/testing/api-mocks/RequestHandlers";
 import { server } from "@/testing/server";
 import { render, screen } from "@/testing/test-utils";
+import { TestUserProvider } from "@/testing/TestUserProvider";
 import { Role } from "@/types/generated/openapi";
 
 import { NavBar } from "./NavBar";
